pyproject.toml
[tool.poetry]
name = "obvcomp"
version = "0.1.0"
description = ""
authors = [
    "Confirm Labs <research@confirmsol.org>",
]

[tool.poetry.dependencies]
python = ">=3.8.1"
typer = "^0.9.0"
tqdm = "^4.65.0"
torch = "^2.0.1"
transformers = "^4.30.2"
datasets = "^2.13.0"
mosaicml-streaming = "^0.5.1"
numpy = "^1.24.3"
pandas = "^2.0.2"
duckdb = "^0.8.1"
boto3 = "^1.26.153"
huggingface-hub = "^0.15.1"
black = "^23.3.0"
ruff = "^0.0.272"
pre-commit = "^3.3.3"

[build-system]
requires = [
    "poetry-core>=1.0.0",
]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
# pyflakes, pycodestyle, isort
select = ["F", "E", "W", "I001"]
exclude = []